+++ src/main/java/com/backendoori/ootw/security/controller/AuthenticationController.java
@@ -2,11 +2,11 @@ package com.backendoori.ootw.security.controller;
 
 import java.util.Optional;
 import com.backendoori.ootw.domain.User;
-import com.backendoori.ootw.security.JwtAuthenticationFilter;
-import com.backendoori.ootw.security.TokenProvider;
 import com.backendoori.ootw.security.dto.LoginDto;
 import com.backendoori.ootw.security.dto.SignupDto;
 import com.backendoori.ootw.security.dto.TokenDto;
+import com.backendoori.ootw.security.jwt.JwtAuthenticationFilter;
+import com.backendoori.ootw.security.jwt.TokenProvider;
 import com.backendoori.ootw.security.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
